Check for file rather than existence
Us os.path.isfile rather than .exists to check for a valid file (not directory).

I have a directory ending in .py that caused a stacktrace in rodeo. A better fix would be around line 24 to exclude directories from the initial list, or check specifically for a directory and report that error, but this should at least prevent the stacktrace.